requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
Flask-Cors==5.0.0
kiteconnect==3.3.0
PyYAML==6.0.1
eventlet==0.33.3
python-dotenv==1.0.1
gunicorn==21.2.0
pytz==2023.3
requests==2.32.3
apscheduler==3.10.4
pyotp==2.9.0
